+++ CHANGELOG.md
@@ -28,6 +28,9 @@ The following methods have been deprecated in favour of new names ([#47](https:/
 - `ZonedDateTime::fromDateTime()` → `fromNativeDateTime()`
 - `ZonedDateTime::toDateTime()` → `toNativeDateTime()`
 - `ZonedDateTime::toDateTimeImmutable()` → `toNativeDateTimeImmutable()`
+- `LocalDateTime::fromDateTime()` → `fromNativeDateTime()`
+- `LocalDateTime::toDateTime()` → `toNativeDateTime()`
+- `LocalDateTime::toDateTimeImmutable()` → `toNativeDateTimeImmutable()`
 
 ## [0.4.0](https://github.com/brick/date-time/releases/tag/0.4.0) - 2021-12-23